-adjusting home page widths
