+++ package.json
@@ -1,5 +1,5 @@
 {
-  "homepage": "http://Krizeon.github.io/krizeon.github.io/",
+  "homepage": "http://Krizeon.github.io/",
   "name": "krizeon.github.io",
   "version": "0.1.0",
   "private": true,
