+++ src/App.css
@@ -6,5 +6,5 @@
   display: block;
   margin-left: auto;
   margin-right: auto;
-  width: 50%;
+  width: 75%;
 }
