+++ src/assets/css/home.css
@@ -1,4 +1,6 @@
 .home {
   background-color: #ffffff;
-  margin: auto;
+  margin-left: auto;
+  margin-right: auto;
+
 }
