+++ src/components/About.js
@@ -1,24 +1,9 @@
 import React from 'react';
 import ContentSeparator from './ContentSeparator'
-import { makeStyles } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
 import Container from '@material-ui/core/Container';
 
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    '& > *': {
-      margin: theme.spacing(1),
-      width: theme.spacing(75),
-      height: theme.spacing(50),
-    },
-  },
-}));
-
 const About = () => {
-  const classes = useStyles();
   return(
     <Container maxWidth="sm">
       <div maxWidth="50">
