+++ src/components/Contact.js
@@ -1,9 +1,12 @@
 import React from 'react';
+import ContentSeparator from './ContentSeparator'
+
 
 const Contact = () => {
   return(
     <>
       <p>This is my Contact Page! Reach me at my email until this page is completed.</p>
+      <ContentSeparator/>
     </>
   );
 }
