+++ src/components/ContentSeparator.js
@@ -4,7 +4,7 @@ import separatorimage from '../assets/images/brush-strokes.png'
 const ContentSeparator = () => {
   return(
     <div>
-      <img class="center" width="20" src={separatorimage} alt="-"/>
+      <img width="200" src={separatorimage} alt="-"/>
     </div>
   );
 }
